+++ packages/client-reconciliation/src/sdk/policiesv1.ts
@@ -208,9 +208,8 @@ export class PoliciesV1 extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }