+++ packages/client-reconciliation/src/sdk/reconciliationsv1.ts
@@ -131,9 +131,8 @@ export class ReconciliationsV1 extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }